# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"\u4f1a\u5458\u5361\u5305API\u6587\u6863","description":"\u8be5\u9879\u76ee\u57fa\u4e8elaravel5.6\u5f00\u53d1","contact":{"name":"Anthony","email":"adshen@google.com"},"version":"1.0"},"servers":[{"url":"http:\/\/card.test\/api","description":"\u672c\u5730\u670d\u52a1\u5668"},{"url":"https:\/\/membercard.adshen.top\/api","description":"\u7ebf\u4e0a\u670d\u52a1\u5668"}],"paths":{"\/card":{"get":{"tags":["\u4f1a\u5458\u5361"],"summary":"\u83b7\u53d6\u6211\u62e5\u6709\u7684\u4f1a\u5458\u5361","description":"\u83b7\u53d6\u767b\u5165\u7528\u6237\u7684\u4f1a\u5458\u5361\u4fe1\u606f","operationId":"createUser","responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"meta":{"properties":{"code":{"type":"integer"},"message":{"type":"string"}},"type":"object"},"data":{"type":"array","items":{}}},"type":"object"}}}}},"security":[{"apiToken":[]}]},"put":{"tags":["\u4f1a\u5458\u5361"],"summary":"\u4fee\u6539\u4f1a\u5458\u5361","operationId":"App\\Http\\Controllers\\CardController::changeCard","requestBody":{"description":"\u4f1a\u5458\u5361\u6570\u636e","required":true,"content":{"application\/json":{"schema":{"properties":{"card_id":{"description":"\u4f1a\u5458\u5361id","type":"integer"},"code":{"description":"\u4f1a\u5458\u5361\u53f7","type":"string","maxLength":30,"minLength":4},"color":{"description":"\u989c\u8272","type":"string"},"info":{"description":"\u5907\u6ce8\u4fe1\u606f","type":"string"}},"type":"object"}}}},"responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"meta":{"properties":{"code":{"type":"integer"},"message":{"type":"string"}},"type":"object"}},"type":"object"}}}}},"security":[{"apiToken":[]}]},"post":{"tags":["\u4f1a\u5458\u5361"],"summary":"\u6dfb\u52a0\u4f1a\u5458\u5361","operationId":"App\\Http\\Controllers\\CardController::addCard","requestBody":{"description":"\u4f1a\u5458\u5361\u6570\u636e","required":true,"content":{"application\/json":{"schema":{"properties":{"code":{"description":"\u4f1a\u5458\u5361\u53f7","type":"string","maxLength":30,"minLength":4},"name":{"description":"\u4f1a\u5458\u5361\u540d\u79f0","type":"string","default":"\u540d\u79f0","maxLength":30,"minLength":2},"type_id":{"description":"\u4f1a\u5458\u5361\u7c7b\u578b\u7f16\u53f7","type":"integer"},"color":{"description":"\u989c\u8272","type":"string"},"info":{"description":"\u5907\u6ce8\u4fe1\u606f","type":"string"},"avatar":{"description":"\u56fe\u6807","type":"string"}},"type":"object"}}}},"responses":{"200":{"description":"\u521b\u5efa\u6210\u529f","content":{"application\/json":{"schema":{"properties":{"meta":{"properties":{"code":{"type":"integer"},"message":{"type":"string"}},"type":"object"}},"type":"object"}}}}},"security":[{"apiToken":[]}]},"delete":{"tags":["\u4f1a\u5458\u5361"],"summary":"\u5220\u9664\u4f1a\u5458\u5361","operationId":"App\\Http\\Controllers\\CardController::deleteCard","requestBody":{"description":"\u4f1a\u5458\u5361id","required":true,"content":{"application\/json":{"schema":{"properties":{"card_id":{"description":"\u4f1a\u5458\u5361id","type":"integer"}},"type":"object"}}}},"responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"meta":{"properties":{"code":{"type":"integer"},"message":{"type":"string"}},"type":"object"}},"type":"object"}}}}},"security":[{"apiToken":[]}]}},"\/card\/type":{"get":{"tags":["\u4f1a\u5458\u5361"],"summary":"\u83b7\u53d6\u4f1a\u5458\u5361\u7c7b\u578b","operationId":"App\\Http\\Controllers\\CardController::getCardType","responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"meta":{"properties":{"code":{"type":"integer"},"message":{"type":"string"}},"type":"object"},"data":{"type":"array","items":{}}},"type":"object"}}}}},"security":[{"apiToken":[]}]}},"\/card\/click":{"post":{"tags":["\u4f1a\u5458\u5361"],"summary":"\u70b9\u51fb\u4f1a\u5458\u5361","operationId":"App\\Http\\Controllers\\CardController::clickCard","requestBody":{"description":"\u4f1a\u5458\u5361id","required":true,"content":{"application\/json":{"schema":{"properties":{"card_id":{"description":"\u4f1a\u5458\u5361id","type":"integer"},"rank":{"description":"\u6392\u5e8f\u7b49\u7ea7","type":"integer"}},"type":"object"}}}},"responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"meta":{"properties":{"code":{"type":"integer"},"message":{"type":"string"}},"type":"object"}},"type":"object"}}}}},"security":[{"apiToken":[]}]}},"\/card\/share":{"get":{"tags":["\u4f1a\u5458\u5361"],"summary":"\u83b7\u53d6\u88ab\u5206\u4eab\u7684\u4f1a\u5458\u5361","operationId":"App\\Http\\Controllers\\CardController::getShareCard","parameters":[{"name":"card_token","in":"query","description":"\u4f1a\u5458\u5361token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"meta":{"properties":{"code":{"type":"integer"},"message":{"type":"string"}},"type":"object"}},"type":"object"}}}}},"security":[{"apiToken":[]}]},"post":{"tags":["\u4f1a\u5458\u5361"],"summary":"\u4f1a\u5458\u5361\u5206\u4eab","operationId":"App\\Http\\Controllers\\CardController::shareCard","requestBody":{"description":"\u4f1a\u5458\u5361id","required":true,"content":{"application\/json":{"schema":{"properties":{"card_id":{"description":"\u4f1a\u5458\u5361id","type":"integer"}},"type":"object"}}}},"responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"meta":{"properties":{"code":{"type":"integer"},"message":{"type":"string"}},"type":"object"}},"type":"object"}}}}},"security":[{"apiToken":[]}]}},"\/card\/global":{"get":{"tags":["\u4f1a\u5458\u5361"],"summary":"\u83b7\u53d6\u5168\u5c40\u4f1a\u5458\u5361","operationId":"App\\Http\\Controllers\\CardController::getGlobalCard","responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"meta":{"properties":{"code":{"type":"integer"},"message":{"type":"string"}},"type":"object"}},"type":"object"}}}}},"security":[{"apiToken":[]}]},"post":{"tags":["\u4f1a\u5458\u5361"],"summary":"\u5168\u5c40\u5206\u4eab\u4f1a\u5458\u5361","operationId":"App\\Http\\Controllers\\CardController::shareCardGlobal","requestBody":{"description":"\u4f1a\u5458\u5361id","required":true,"content":{"application\/json":{"schema":{"properties":{"card_id":{"description":"\u4f1a\u5458\u5361id","type":"integer"}},"type":"object"}}}},"responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"meta":{"properties":{"code":{"type":"integer"},"message":{"type":"string"}},"type":"object"}},"type":"object"}}}}},"security":[{"apiToken":[]}]}}},"components":{"schemas":{"User":{}},"securitySchemes":{"apiToken":{"type":"apiKey","description":"\u6388\u6743Token","name":"X-AUTH-TOKEN","in":"header"}}},"tags":[{"name":"\u4f1a\u5458\u5361","description":"\u4f1a\u5458\u5361\u76f8\u5173API"}]}